refactor: replace useTokenLists with useTokenLists2
